Add debouncing to ParsedInput
It feels bad that it took this much refactoring and new code to add
something as common as input debouncing; I guess that's the cost of a
pure functional language. While it works, it still feels messy and hard
to understand and I'm not sure how to make it better.

Also disables the subtraction input until I can convert it to the new
ParsedInput API.